Attempt at css transition
- It doesn't work great because the only way to hide the divs is
with either overflow-x: hidden, which adds a vertical scroll bar
to any parent, or transitioning width - and I can't figure out
how to transition width so it starts on the right and opens up
to the left (at least with just css). Also, transitioning with
width makes things look weird. Going to go back to js.