Removed comments about moving mouseover outside of function
- Reading Source (3)'s comments more closely: this is reset on
every update, so it will use the latest chordPaths selection
This means that it does have to be redone in the function. Not
sure how to keep the queued events from piling up, but it is
necessary and hopefully people will mouse over frequently enough
for it not to be an issue.